Refactor media analysis integration to use FfProbeData
- Replaced instances of IMediaAnalysis with FfProbeData in MusicLogic, FFmpegCommandBuilder, VideoAudioFile, and related classes.
- Updated methods to accommodate new FfProbeData structure, ensuring proper handling of audio, video, and subtitle streams.
- Enhanced error handling and process management in Shell.cs to ensure child processes are terminated correctly.
- Modified metadata handling in FileManager and FileRepository to utilize FfProbeData for improved performance and accuracy.
- Added functionality to delete empty playlists in the Dev class for better media management.

+++ src/NoMercy.Database/Models/Metadata.cs
@@ -166,9 +166,9 @@ public class IAudio : IHash
     [JsonProperty("language")] public string Language { get; set; } = null!;
     [JsonProperty("codec")] public string? Codec { get; set; }
     [JsonProperty("bit_rate")] public long? BitRate { get; set; }
-    [JsonProperty("channels")] public int? Channels { get; set; }
+    [JsonProperty("channels")] public long? Channels { get; set; }
     [JsonProperty("channel_layout")] public string? ChannelLayout { get; set; }
-    [JsonProperty("sample_rate")] public int? SampleRate { get; set; }
+    [JsonProperty("sample_rate")] public long? SampleRate { get; set; }
 }
 
 public class ISubtitle : IHash
